+++ GR_ChangeLog.lua
@@ -7,6 +7,7 @@ ns.changeLog = [[
     ### v4.0.11
         - Created the about page.
         - Added fix for guild name change.
+        - Implemented Analytics (can only see session for now).
     ### v4.0.10
         - Added What's New Screen
         - Added ability to change send message delay.